% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D8Flowdir.R
\name{D8Flowdir}
\alias{D8Flowdir}
\title{compute the D8 flow directions using a digital elevation model (DEM).}
\usage{
D8Flowdir(dem, out_dir = getwd(), slope_prefix = NULL,
  fldir_prefix = NULL, slope_suffix = "sd8", fldir_suffix = "p")
}
\arguments{
\item{dem}{character path to the raster file, A digital elevation model.}

\item{out_dir}{character, the path to the directory where the output file is meant to be saved.}

\item{slope_prefix}{character (optional), the choosen base file name for slope output file.}

\item{fldir_prefix}{character (optional), the choosen base file name for flow directions output file.}

\item{slope_suffix}{character (optional), suffix to be added to the slope output file, useful for tracking purpose.}

\item{fldir_suffix}{character (optional), suffix to be added to the flow directions output file, useful for tracking purpose.}
}
\value{
raster object
}
\description{
Take a DEM (typically the depression filled file returned by Pitremove),
and return a raster files containing the D8 slope and the D8 flow directions.
}
\details{
TauDEM version 5.3 has been developed to read any raster supported by GDAL
and write grid files in the GeoTiff (.tif) format only. TauDEM reads and writes ESRI
shape files and text files generally identified using the .txt extension.
}
\examples{
# This may take time and requires a connection to internet to download a DEM of size 5.3 MB.
destfile <- "C:/Users/LIssoufou/Downloads/LoganDemo.zip"
exdir <- "C:/Users/LIssoufou/Downloads"
exdir <- paste0 (exdir, "/", "LoganDemo")
dem <- paste0 (exdir, "/","Logan", "/", "logan.tif")
download.file(url = "http://hydrology.usu.edu/taudem/taudem5/LoganDemo.zip", destfile = destfile)
unzip(zipfile = destfile, exdir = exdir)
library(rgdal)
library(raster)
z <- raster(dem)
plot(z)
out_dir <- paste0(exdir, "/", "R_examples")
dir.create(file.path(exdir, 'R_examples'))
prefix <- 'logan'
suffix <- 'fel'
fel <- Pitremove (dem = dem, out_dir = out_dir, prefix = prefix, suffix = suffix)
fel <- raster(paste0(out_dir,"/", prefix, suffix, ".tif"))
plot(fel)
dem <- paste0(out_dir,"/", prefix, suffix, ".tif")
D8Flowdir(dem=dem, out_dir = out_dir, slope_prefix = "slope_logan", fldir_prefix = "fldir_logan",
slope_suffix = "sd8", fldir_suffix="p")
}
