% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pitremove.R
\name{Pitremove}
\alias{Pitremove}
\title{remove pits by filling depressions in a digital elevation model (DEM).}
\usage{
Pitremove(dem, out_dir = getwd(), prefix = NULL, suffix = "fel")
}
\arguments{
\item{dem}{character path to the raster file, A digital elevation model.}

\item{out_dir}{character, the path to the directory where the output file is meant to be saved.}

\item{prefix}{character (optional), the choosen base file name.}

\item{suffix}{character (optional), suffix to be added to the output file, useful for tracking
purpose. default is "fel" as by TauDEM file naming convention.}
}
\value{
raster object
}
\description{
Take a DEM, and a return a file comprising the DEM with pits removed.
}
\details{
TauDEM version 5.3 has been developed to read any raster supported by GDAL
and write grid files in the GeoTiff (.tif) format only. TauDEM reads and writes ESRI
shape files and text files generally identified using the .txt extension.
}
\examples{
# This may take time and requires a connection to internet to download a DEM of size 5.3 MB.
destfile <- "C:/Users/LIssoufou/Downloads/LoganDemo.zip"
exdir <- "C:/Users/LIssoufou/Downloads"
exdir <- paste0 (exdir, "/", "LoganDemo")
dem <- paste0 (exdir, "/","Logan", "/", "logan.tif")
download.file(url = "http://hydrology.usu.edu/taudem/taudem5/LoganDemo.zip", destfile = destfile)
unzip(zipfile = destfile, exdir = exdir)
library(rgdal)
library(raster)
z <- raster(dem)
plot(z)
out_dir <- paste0(exdir, "/", "R_examples")
dir.create(file.path(exdir, 'R_examples'))
prefix <- 'logan'
suffix <- 'fel'
fel <- Pitremove (dem = dem, out_dir = out_dir, prefix = prefix, suffix = suffix)
fel <- raster(paste0(out_dir,"/", prefix, suffix, ".tif"))
plot(fel)
}
