% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D8Flowdir.R
\name{D8Flowdir}
\alias{D8Flowdir}
\title{compute the D8 flow directions using a digital elevation model (DEM).}
\usage{
D8Flowdir(fun = "D8Flowdir", felfile, sd8file = NULL, pfile = NULL,
  number_of_processes = 8)
}
\arguments{
\item{fun}{the name of flow direction function depending on the version TauDEM in use.}

\item{felfile}{character, path to the raster file, A digital elevation model (Must be TIFF File and typical
submited to pitremove function aprior).}

\item{sd8file}{character, path to the output slope file to be computed along the way by D8Flowdir.}

\item{pfile}{character, path to the output flow direction file to be computed as end product.}

\item{number_of_processes}{number of MPI processes to be used. The default is 8.This is the
number of stripes that the domain will be divided into and the number of MPI parallel processes
that will be spawned to evaluate each of the stripes.}
}
\value{
raster layer
}
\description{
Take a DEM (typically the depression filled file returned by Pitremove),
and return a raster files containing the D8 slope and the D8 flow directions.
}
\details{
TauDEM version 5.3 has been developed to read any raster supported by GDAL
and write grid files in the GeoTiff (.tif) format only. TauDEM reads and writes ESRI
shape files and text files generally identified using the .txt extension.

D8Flowdir creates 2 grids. The first contains the flow direction from each grid cell to one of its adjacent
or diagonal neighbors, calculated using the direction of steepest descent. The second contain the slope, as
evaluated in the direction of steepest descent, and is reported as drop/distance, i.e. tan of the angle.
Flow direction is reported as "no data" for any grid cell adjacent to the edge of the DEM domain, or adjacent
to a "no data" value in the DEM. In flat areas, flow directions are assigned away from higher ground and
towards lower ground using the method of Garbrecht and Martz (1997). The D8 flow direction algorithm may be
applied to a DEM that has not had its pits filled, but it will then result in "no data" values for flow
direction and slope at the lowest point of each pit.
}
