% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDEM.R
\name{getDEM}
\alias{getDEM}
\title{Download digital elevation model (DEM) by shapefile}
\usage{
getDEM(shp, path = "")
}
\arguments{
\item{shp}{SpatialPolygonsDataFrame shapefile from the sp package}

\item{path}{Character. Path name indicating where to store the data. Default is the current working directory }
}
\value{
a rater layer or a list of raster layers (if the shapefile covers several tiles).
}
\description{
Takes a shapefile and download SRTM DEM covering the shapefile
}
